Add dependency: Amazon Kinesis Client Library for Java
This commit adds dependency to most recent release of
com.amazonaws:amazon-kinesis-client

As part of this, we exclude commons-logging, because we are following
Spring Boot best practices which favors logging to the slf4j interface
and use of the Logback implementation.

== Add dependency: Amazon Kinesis Client Library for Java

See:
https://github.com/awslabs/amazon-kinesis-client

Latest release as of Apr 10, 2017 is 1.7.5

== Exclude commons-logging

Exclude commons-logging using "alternative 3" technique described here:
https://www.slf4j.org/faq.html#excludingJCL

This app uses the SLF4J logging interface, and the Logback implementation.

Any dependencies that depend on commons-logging API will use implementation
provided by jcl-over-slf4j.

All the Spring Boot starters depend on spring-boot-starter-logging which, in
turn, depends on logback-classic and jcl-over-slf4j.

Learn more:
https://docs.spring.io/spring-boot/docs/current/reference/html/howto-logging.html